Introduce serialNsupport for variationo of number of serial ports.
This change allows a device to specify the number of serial port
it supports, and changes the code to use one of serial1support,
serial2support,serial3support defines to declare support for the
number of serial ports it supports.

This avoids having the "#if defined(...)" for every device which
may support the additional serial port(s).