Fix testescape() waiting a long time for input
When Serial is a CDC-ACM, the Serial.read() function blocks.  This meant that
everytime testescape() was called, it would block indefinitely, until a
character was received.  This change tests to see if anything can be read
before attempting to read.

With this change, the "blink" example works (most times):
   (defun b (pin ms)
     (pinmode pin t)
     (loop
       (digitalwrite pin t)
       (delay ms)
       (digitalwrite pin nil)
       (delay ms)))

Issues still remain with sporadic input (especially at the start) ...